psx rewrite began

+++ src/psx/decode_xa.cpp
@@ -21,7 +21,6 @@
 * XA audio decoding functions (Kazzuya).
 */
 
-#include "psxcommon.h"
 #include "decode_xa.h"
 
 #define FIXED
@@ -341,7 +340,7 @@ static int parse_xa_audio_sector( xa_decode_t *xdp,
 //===                  - 0 for any other successive sector
 //=== return -1 if error
 //================================================================
-long xa_decode_sector( xa_decode_t *xdp,
+s32 xa_decode_sector( xa_decode_t *xdp,
 					   unsigned char *sectorp, int is_first_sector ) {
 	if (parse_xa_audio_sector(xdp, (xa_subheader_t *)sectorp, sectorp + sizeof(xa_subheader_t), is_first_sector))
 		return -1;
